+++ lib/Migrator/APIDiffMigratorPass.cpp
@@ -1,4 +1,4 @@
-//===--- SyntacticMigratorPass.cpp ----------------------------------------===//
+//===--- APIDiffMigratorPass.cpp ------------------------------------------===//
 //
 // This source file is part of the Swift.org open source project
 //
@@ -10,17 +10,16 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "swift/AST/NameLookup.h"
 #include "swift/AST/USRGeneration.h"
 #include "swift/AST/ASTVisitor.h"
 #include "swift/Frontend/Frontend.h"
 #include "swift/IDE/Utils.h"
 #include "swift/Index/Utils.h"
+#include "swift/Migrator/ASTMigratorPass.h"
 #include "swift/Migrator/EditorAdapter.h"
 #include "swift/Migrator/FixitApplyDiagnosticConsumer.h"
 #include "swift/Migrator/Migrator.h"
 #include "swift/Migrator/RewriteBufferEditsReceiver.h"
-#include "swift/Migrator/SyntacticMigratorPass.h"
 #include "clang/Basic/Diagnostic.h"
 #include "clang/Basic/FileManager.h"
 #include "clang/Basic/SourceManager.h"
@@ -33,6 +32,8 @@ using namespace swift;
 using namespace swift::migrator;
 using namespace swift::ide;
 using namespace swift::ide::api;
+
+namespace {
 
 struct FoundResult {
   SourceRange TokenRange;
@@ -58,7 +59,8 @@ public:
         bool Optional = End.isValid();
         if (!Optional)
           End = Init->getNameLoc();
-        return {SourceRange(Init->getNameLoc(), End), Optional, /*suffixable=*/true};
+        return {SourceRange(Init->getNameLoc(), End), Optional,
+          /*suffixable=*/true};
       }
       return {SourceRange(), false, false};
     }
@@ -205,13 +207,7 @@ public:
   }
 };
 
-struct SyntacticMigratorPass::Implementation : public SourceEntityWalker {
-  SourceFile *SF;
-  const StringRef FileName;
-  unsigned BufferId;
-  SourceManager &SM;
-  EditorAdapter &Editor;
-  const MigratorOptions &Opts;
+struct APIDiffMigratorPass : public ASTMigratorPass, public SourceEntityWalker {
 
   APIDiffItemStore DiffStore;
 
@@ -273,10 +269,9 @@ struct SyntacticMigratorPass::Implementation : public SourceEntityWalker {
     return false;
   }
 
-  Implementation(SourceFile *SF, EditorAdapter &Editor,
-                 const MigratorOptions &Opts) :
-    SF(SF), FileName(SF->getFilename()), BufferId(SF->getBufferID().getValue()),
-      SM(SF->getASTContext().SourceMgr), Editor(Editor), Opts(Opts) {}
+  APIDiffMigratorPass(EditorAdapter &Editor, SourceFile *SF,
+                      const MigratorOptions &Opts)
+    : ASTMigratorPass(Editor, SF, Opts) {}
 
   void run() {
     if (Opts.APIDigesterDataStorePath.empty())
@@ -368,210 +363,8 @@ struct SyntacticMigratorPass::Implementation : public SourceEntityWalker {
     }
   }
 
-  /// Migrates code that compiles fine in Swift 3 but breaks in Swift 4 due to
-  /// changes in how the typechecker handles tuple arguments.
-  void handleTupleArgumentMismatches(const CallExpr *E) {
-    if (!SF->getASTContext().LangOpts.isSwiftVersion3())
-      return;
-    if (E->isImplicit())
-      return;
-
-    // Handles such kind of cases:
-    // \code
-    //   func test(_: ()) {}
-    //   test()
-    // \endcode
-    // This compiles fine in Swift 3 but Swift 4 complains with
-    //   error: missing argument for parameter #1 in call
-    //
-    // It will fix the code to "test(())".
-    //
-    auto handleCallsToEmptyTuple = [&](const CallExpr *E) -> bool {
-      auto fnTy = E->getFn()->getType()->getAs<FunctionType>();
-      if (!fnTy)
-        return false;
-      auto parenT = dyn_cast<ParenType>(fnTy->getInput().getPointer());
-      if (!parenT)
-        return false;
-      auto inp = dyn_cast<TupleType>(parenT->getUnderlyingType().getPointer());
-      if (!inp)
-        return false;
-      if (inp->getNumElements() != 0)
-        return false;
-      auto argTupleT = dyn_cast<TupleType>(E->getArg()->getType().getPointer());
-      if (!argTupleT)
-        return false;
-      if (argTupleT->getNumElements() != 0)
-        return false;
-      Editor.insertWrap("(", E->getArg()->getSourceRange(), ")");
-      return true;
-    };
-
-    // Handles such kind of cases:
-    // \code
-    //   func test(_: ((Int, Int)) -> ()) {}
-    //   test({ (x,y) in })
-    // \endcode
-    // This compiles fine in Swift 3 but Swift 4 complains with
-    //   error: cannot convert value of type '(_, _) -> ()' to expected argument type '((Int, Int)) -> ()'
-    //
-    // It will fix the code to "test({ let (x,y) = $0; })".
-    //
-    auto handleTupleMapToClosureArgs = [&](const CallExpr *E) -> bool {
-      auto fnTy = E->getFn()->getType()->getAs<FunctionType>();
-      if (!fnTy)
-        return false;
-      auto fnTy2 = fnTy->getInput()->getAs<FunctionType>();
-      if (!fnTy2)
-        return false;
-      auto parenT = dyn_cast<ParenType>(fnTy2->getInput().getPointer());
-      if (!parenT)
-        return false;
-      auto tupleInFn = dyn_cast<TupleType>(parenT->getUnderlyingType().getPointer());
-      if (!tupleInFn)
-        return false;
-      if (!E->getArg())
-        return false;
-      auto argE = E->getArg()->getSemanticsProvidingExpr();
-      while (auto *ICE = dyn_cast<ImplicitConversionExpr>(argE))
-        argE = ICE->getSubExpr();
-      argE = argE->getSemanticsProvidingExpr();
-      auto closureE = dyn_cast<ClosureExpr>(argE);
-      if (!closureE)
-        return false;
-      if (closureE->getInLoc().isInvalid())
-        return false;
-      auto paramList = closureE->getParameters();
-      if (!paramList ||
-          paramList->getLParenLoc().isInvalid() || paramList->getRParenLoc().isInvalid())
-        return false;
-      if (paramList->size() != tupleInFn->getNumElements())
-        return false;
-      if (paramList->size() == 0)
-        return false;
-
-      auto hasParamListWithNoTypes = [&]() {
-        if (closureE->hasExplicitResultType())
-          return false;
-        for (auto *param : *paramList) {
-          auto tyLoc = param->getTypeLoc();
-          if (!tyLoc.isNull())
-            return false;
-        }
-        return true;
-      };
-
-      if (hasParamListWithNoTypes()) {
-        // Simpler form depending on type inference.
-        // Change "(x, y) in " to "let (x, y) = $0;".
-
-        Editor.insert(paramList->getLParenLoc(), "let ");
-        for (auto *param : *paramList) {
-          // If the argument list is like "(_ x, _ y)", remove the underscores.
-          if (param->getArgumentNameLoc().isValid()) {
-            Editor.remove(CharSourceRange(SM, param->getArgumentNameLoc(),
-                                          param->getNameLoc()));
-          }
-          // If the argument list has type annotations, remove them.
-          auto tyLoc = param->getTypeLoc();
-          if (!tyLoc.isNull() && !tyLoc.getSourceRange().isInvalid()) {
-            auto nameRange = CharSourceRange(param->getNameLoc(),
-                                             param->getNameStr().size());
-            auto tyRange = Lexer::getCharSourceRangeFromSourceRange(SM,
-                                                        tyLoc.getSourceRange());
-            Editor.remove(CharSourceRange(SM, nameRange.getEnd(),
-                                          tyRange.getEnd()));
-          }
-        }
-
-        Editor.replaceToken(closureE->getInLoc(), "= $0;");
-        return true;
-      }
-
-      // Includes types in the closure signature. The following will do a
-      // more complicated edit than the above:
-      //   (x: Int, y: Int) -> Int in
-      // to
-      //   (__val:(Int, Int)) -> Int in let (x,y) = __val;
-
-      std::string paramListText;
-      {
-        llvm::raw_string_ostream OS(paramListText);
-        OS << "(__val:(";
-        for (size_t i = 0, e = paramList->size(); i != e; ++i) {
-          if (i != 0)
-            OS << ", ";
-          auto param = paramList->get(i);
-          auto tyLoc = param->getTypeLoc();
-          if (!tyLoc.isNull() && !tyLoc.getSourceRange().isInvalid()) {
-            OS << SM.extractText(
-              Lexer::getCharSourceRangeFromSourceRange(SM,
-                                                       tyLoc.getSourceRange()));
-          } else {
-            param->getType().print(OS);
-          }
-        }
-        OS << "))";
-      }
-      std::string varBindText;
-      {
-        llvm::raw_string_ostream OS(varBindText);
-        OS << " let (";
-        for (size_t i = 0, e = paramList->size(); i != e; ++i) {
-          if (i != 0)
-            OS << ",";
-          auto param = paramList->get(i);
-          OS << param->getNameStr();
-        }
-        OS << ") = __val; ";
-      }
-
-      Editor.replace(paramList->getSourceRange(), paramListText);
-      Editor.insertAfterToken(closureE->getInLoc(), varBindText);
-      return true;
-    };
-
-    if (handleCallsToEmptyTuple(E))
-      return;
-    if (handleTupleMapToClosureArgs(E))
-      return;
-  }
-
-  /// If the expression is `type(of:)`, in Swift 3 it's a special 'dynamic type
-  /// expression'. In Swift 4, this is resolved by overload resolution, so
-  /// if there is a func type(of:) or even var type, it can shadow
-  /// Swift.type(of:), which is probably not what people want.
-  ///
-  /// This migration operation adds `Swift.` to disambiguate in Swift 4.
-  void handleTypeOfFunction(const DynamicTypeExpr *DTE) {
-    if (!SF->getASTContext().LangOpts.isSwiftVersion3()) {
-      return;
-    }
-    const auto CSR = Lexer::getCharSourceRangeFromSourceRange(SM,
-      DTE->getSourceRange());
-    if (SM.extractText(CSR).startswith("Swift.")) {
-      return;
-    }
-
-    UnqualifiedLookup Lookup {
-      { SF->getASTContext().getIdentifier("type") },
-      SF->getModuleScopeContext(),
-      /*TypeResolver=*/nullptr,
-      /*IsKnownPrivate=*/false,
-      DTE->getLoc()
-    };
-    if (Lookup.Results.empty()) {
-      // There won't be a name shadowing here in Swift 4, so we don't need to
-      // do anything.
-      return;
-    }
-
-    Editor.insertBefore(DTE->getLoc(), "Swift.");
-  }
-
   bool walkToExprPre(Expr *E) override {
     if (auto *CE = dyn_cast<CallExpr>(E)) {
-      handleTupleArgumentMismatches(CE);
 
       auto Fn = CE->getFn();
       auto Args = CE->getArg();
@@ -599,11 +392,6 @@ struct SyntacticMigratorPass::Implementation : public SourceEntityWalker {
         break;
       }
     }
-
-    if (const auto *DTE = dyn_cast<DynamicTypeExpr>(E)) {
-      handleTypeOfFunction(DTE);
-    }
-
     return true;
   }
 
@@ -688,14 +476,10 @@ struct SyntacticMigratorPass::Implementation : public SourceEntityWalker {
   }
 };
 
-SyntacticMigratorPass::
-SyntacticMigratorPass(EditorAdapter &Editor, SourceFile *SF,
-  const MigratorOptions &Opts) : Impl(*new Implementation(SF, Editor, Opts)) {}
+} // end anonymous namespace
 
-SyntacticMigratorPass::~SyntacticMigratorPass() { delete &Impl; }
-
-void SyntacticMigratorPass::run() { Impl.run(); }
-
-const clang::edit::Commit &SyntacticMigratorPass::getEdits() const {
-  return Impl.Editor.getEdits();
+void migrator::runAPIDiffMigratorPass(EditorAdapter &Editor,
+                                      SourceFile *SF,
+                                      const MigratorOptions &Opts) {
+  APIDiffMigratorPass { Editor, SF, Opts }.run();
 }
