+++ test/Migrator/tuple-arguments.swift
@@ -1,6 +1,3 @@
-// rdar://problem/32025974
-// XFAIL: linux
-
 // RUN: %target-swift-frontend -typecheck %s -swift-version 3
 // RUN: %target-swift-frontend -typecheck -update-code -primary-file %s -emit-migrated-file-path %t.result -disable-migrator-fixits -swift-version 3
 // RUN: diff -u %s.expected %t.result
